requirements.txt
cvxpy==1.21.5
numpy==1.1.15
tqdm==4.64.1
matplotlib==3.5.2
pycddlib==2.1.6
